requirements.txt
# Fichier requirements.txt (CORRIGÉ)

# Pour Apache Airflow et le provider PostgreSQL
apache-airflow-providers-postgres

# Pour appeler l'API
requests

# Pour se connecter à PostgreSQL depuis Python
psycopg2-binary

# Pour l'analyse de sentiment
textblob

# BONNE PRATIQUE : Installe dbt-postgres ET la bonne version de dbt-core.
# Supprimez les lignes pour "dbt-core" et "dbt-postgres" et remplacez-les par celle-ci.
dbt-postgres==1.8.2